caldisplay/server/requirements.txt
ics~=0.7.2
requests~=2.28.2
flask~=2.2.2
pillow~=9.4.0
